Initial update for Making Embedded Systems project
Invent an Imaginary Embedded System

Tell us about an embedded system that does not exist. It can be grandiose or strange; it can be
something just one step beyond what you have around you. Ideally it is something you
personally want to build (even if physics doesn’t allow it). What is it? What does it do? What
general components do you need?

This should be summarized in 3-6 sentences. A diagram would be nice but not required.

Write your idea in a Google doc or a md file on Github and submit your assignment link to the
#assignment-submission channel on Discord. Make sure your submission is publicly viewable.